requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
plotly>=5.0.0
Pillow
opencv-python
google-genai
openai
